moved logic to modules
moved logic to modules
