+++ maplib/mapkeys.js
@@ -1,0 +1,7 @@
+module.exports = {
+
+    signingKeys: ['IFEesFMESV(sV(M(#M$R#@843242(*', 'F(s9fMSEf39394329$M324#@$', 'DFsd9VsD(VM(3(M332424'],
+    jwtAlogrithm: 'HS512',
+    jwtSecretSalt: 'F(SEFsjv*SZV*XZV8efes8f',
+    verboseErrors: true
+};
